Add files via upload
This R script selects significant associations between phenotypes and variants. It also performs fine mapping as a final stage of analysis. It takes as input the files included in the folder UKBB, available in this repository. Just put the script within the folder and run it. No need to unzip the files included in UKBB.